requirements.txt
aioredis==1.3.1
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.2
Automat==20.2.0
cffi==1.14.0
click==7.1.1
constantly==15.1.0
cryptography==2.9
daphne==2.5.0
Django==3.0.5
djangorestframework==3.11.0
h11==0.9.0
hiredis==1.0.1
httptools==0.1.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
msgpack==0.6.2
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2019.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
zope.interface==5.1.0
